move test to workspace.rs, as the `toggle_theme_mode` is inside there.
